UI: Move the Markdown stuff out into its own component.
