+++ packages/ui/src/routes/HomePage.tsx
@@ -1,24 +1,14 @@
 import {faCaretDown, faCaretRight} from "@fortawesome/free-solid-svg-icons";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {MDXProvider} from "@mdx-js/react";
 import * as React from "react";
 import styled from "styled-components";
-import Readme from "../../../../README.mdx";
 import ComponentList from "../components/ComponentList";
-import Highlighter from "../components/Highlighter";
+import {Readme} from "../components/Markdown";
 import {Context} from "../models";
 
 interface Props {
   context: Context;
 }
-
-const StyledMarkdown = styled.div`
-  max-width: 70ch;
-  line-height: 1.4;
-  a {
-    color: ${({theme}) => theme.colors.textSecondary};
-  }
-`;
 
 const StyledDocsToggle = styled.div`
   font-size: ${({theme}) => theme.fontSizes.subtitle};
@@ -29,7 +19,6 @@ const HomePage = ({context}: Props) => {
     context.components.length === 0,
   );
   const hasComponents = context.components.length > 0;
-  const mdxComponents = {code: Highlighter};
 
   return (
     <>
@@ -40,11 +29,7 @@ const HomePage = ({context}: Props) => {
               <FontAwesomeIcon icon={faCaretDown} /> Hide documentation
             </StyledDocsToggle>
           )}
-          <MDXProvider components={mdxComponents}>
-            <StyledMarkdown>
-              <Readme />
-            </StyledMarkdown>
-          </MDXProvider>
+          <Readme />
         </React.Fragment>
       ) : (
         hasComponents && (
